[IMP] base_delivery_carrier_label: Separate test models
