--- base_delivery_carrier_label/tests/models/__init__.py
@@ -1,1 +1,0 @@
-from . import carrier_test
